update install scripts. TODO call update script in install

+++ dwh-install/src/main/scripts/install_centos.sh
@@ -1,22 +1,26 @@
 #!/usr/bin/env bash
 
 SCRIPT=$(readlink -f "$0")
-install_root=$(dirname "$SCRIPT")/
+INSTALL_ROOT=$(dirname "$SCRIPT")/
 
-MY_PATH=$install_root
-
-DATA_HOME=$MY_PATH/i2b2_install
-DATA_DEST=$MY_PATH/temp_install
-PACKAGES=$MY_PATH/packages
+DATA_HOME=$INSTALL_ROOT/i2b2_install
+DATA_DEST=$INSTALL_ROOT/temp_install
+PACKAGES=$INSTALL_ROOT/packages
 
 WILDFLY_HOME=/opt/wildfly-9.0.2.Final
 JBOSS7_DIR=/opt/jboss-as-7.1.1.Final 
 
+echo
+echo +++++ STEP 0 +++++ Installation der notwendigen Pakete | tee -a $LOGFILE
+echo
 yum clean all
 yum -y update
 yum -y install java-1.8.0-openjdk-headless
 yum -y install sudo wget curl dos2unix unzip sed bc ant php php-curl openssh-server
 
+echo
+echo +++++ STEP 0.01 +++++ CentOs HTTPD Verlinkungen | tee -a $LOGFILE
+echo
 # make centos preparations
 ln -s /etc/httpd /etc/apache2
 mkdir /etc/httpd/conf-available
@@ -24,6 +28,7 @@ mkdir /etc/httpd/conf-enabled
 echo IncludeOptional conf-enabled/*.conf >> /etc/httpd/conf/httpd.conf
 
 # reverse proxy configuration
+echo Reverse Proxy Konfigurierung | tee -a $LOGFILE
 conf=/etc/apache2/conf-available/aktin-j2ee-reverse-proxy.conf
 echo ProxyPreserveHost On > $conf
 echo ProxyPass /aktin http://localhost:9090/aktin >> $conf
@@ -35,125 +40,166 @@ echo LoadModule proxy_module libexec/apache2/mod_proxy.so >> /etc/httpd/conf/htt
 echo LoadModule proxy_http_module libexec/apache2/mod_proxy_http.so >> /etc/httpd/conf/httpd.conf
 
 #apachectl restart
-systemctl enable httpd
-systemctl start httpd
+echo HTTPD ins Autostart übernehmen und starten | tee -a $LOGFILE
+systemctl enable httpd 2>&1 | tee -a $LOGFILE
+systemctl start httpd 2>&1 | tee -a $LOGFILE
 
 # open port 80
-firewall-cmd --zone=public --add-port=80/tcp --permanent
-firewall-cmd --reload
+echo Firewall Konfigurierung | tee -a $LOGFILE
+firewall-cmd --zone=public --add-port=80/tcp --permanent 2>&1 | tee -a $LOGFILE
+firewall-cmd --reload 2>&1 | tee -a $LOGFILE
 
+echo linking Documentroot /var/www/html to /var/webroot | tee -a $LOGFILE
 ln -s /var/www/html /var/webroot
 
-#ln -s $install_root /opt/aktin
+#ln -s $INSTALL_ROOT /opt/aktin
 
-#install_root=/opt/aktin
+#INSTALL_ROOT=/opt/aktin
 
+echo
+echo +++++ STEP 0.02 +++++ Postgres Konfigurationen | tee -a $LOGFILE
+echo
 #postgres
 yum -y install postgresql-server postgresql-contrib
-postgresql-setup initdb
-systemctl enable postgresql
+postgresql-setup initdb 2>&1 | tee -a $LOGFILE
+systemctl enable postgresql 2>&1 | tee -a $LOGFILE
 
 sudo -u postgres cp /var/lib/pgsql/data/pg_hba.conf /var/lib/pgsql/data/pg_hba.conf.orig
 cat /var/lib/pgsql/data/pg_hba.conf.orig | sudo -u postgres sed -r -e 's|(host\W+all\W+all\W+127.0.0.1/32\W+)ident|\1trust|' -e's|(host\W+all\W+all\W+::1/128\W+)ident|\1trust|' -e 's|(local\W+all\W+all\W+)peer|\1trust|' > /var/lib/pgsql/data/pg_hba.conf
 
-systemctl start postgresql
+systemctl start postgresql 2>&1 | tee -a $LOGFILE
 
-sudo cp /etc/sysconfig/selinux $install_root/selinux.orig
-sudo cat $install_root/selinux.orig | sudo sed 's|SELINUX=enforcing|SELINUX=disabled|' > /etc/sysconfig/selinux
 
-LOG_DIR=$install_root/logs
+echo
+echo +++++ STEP 0.03 +++++ R Konfigurationen | tee -a $LOGFILE
+echo
+# install R libraries for reporting, adding fedora repos
+echo Über Fedora Repo R beziehen : | tee -a $LOGFILE
+rpm -Uvh http://download.fedoraproject.org/pub/epel/7/x86_64/e/epel-release-7-9.noarch.rpm 2>&1 | tee -a $LOGFILE
+# preparations for r-xml
+yum -y install libxml2 libxml2-devel 2>&1 | tee -a $LOGFILE
+yum -y install R  2>&1 | tee -a $LOGFILE
+Rscript -e 'install.packages("XML", repos="https://cran.rstudio.com/")'  2>&1 | tee -a $LOGFILE
+Rscript -e 'install.packages("lattice", repos="https://cran.rstudio.com/")' 2>&1 | tee -a $LOGFILE
+
+
+echo
+echo +++++ STEP 0.04 +++++ Deactivierung von SELinux (restart required) | tee -a $LOGFILE
+echo
+sudo cp /etc/sysconfig/selinux $INSTALL_ROOT/selinux.orig
+sudo cat $INSTALL_ROOT/selinux.orig | sudo sed 's|SELINUX=enforcing|SELINUX=disabled|' > /etc/sysconfig/selinux
+
+
+
+echo
+echo +++++ STEP 1 +++++ Prüfung aktin.properties und email.config | tee -a $LOGFILE
+echo
+count=0
+if [ ! -f $INSTALL_ROOT/lib/aktin.properties ] ; then
+	echo Bitte die Datei $INSTALL_ROOT/aktin.properties mit Ihren Daten füllen und nach $INSTALL_ROOT/lib/ kopieren : | tee -a $LOGFILE 
+	echo nano $INSTALL_ROOT/aktin.properties | tee -a $LOGFILE
+	echo cp $INSTALL_ROOT/aktin.properties $INSTALL_ROOT/lib/ | tee -a $LOGFILE
+	echo  | tee -a $LOGFILE
+	count=1
+fi
+if [ ! -f $INSTALL_ROOT/lib/email.config ] ; then
+	echo Bitte die Datei $INSTALL_ROOT/email.config mit Ihren Daten füllen und nach $INSTALL_ROOT/lib/ kopieren :  | tee -a $LOGFILE
+	echo nano $INSTALL_ROOT/email.config | tee -a $LOGFILE
+	echo cp $INSTALL_ROOT/email.config $INSTALL_ROOT/lib/ | tee -a $LOGFILE
+	echo  | tee -a $LOGFILE
+	count=1
+fi
+if [ $count -gt 0 ]; then
+	RCol='\e[0m'
+	Red='\e[0;31m'
+	echo -e "${Red} Das Skript wird jetzt beendet. Bitte rufen Sie es erneut auf, wenn Sie die notwendigen Änderungen durchgeführt haben${RCol}" | tee -a $LOGFILE
+	echo  | tee -a $LOGFILE
+	exit 1
+fi
+
+
+echo
+echo +++++ STEP 2 +++++ Links und Rechte | tee -a $LOGFILE
+echo
+
+chmod -R o+x $INSTALL_ROOT
+
+echo
+echo +++++ STEP 2.01 +++++ Log Ordner | tee -a $LOGFILE
+echo
+# create directory for logs if not existent
+# TODO dont write logfiles to /vagrant
+LOG_DIR=$INSTALL_ROOT/logs
 if [ ! -d "$LOG_DIR" ]; then 
     mkdir -p $LOG_DIR
     chmod -R 777 $LOG_DIR
 fi
 
-# install R libraries for reporting, adding fedora repos
-rpm -Uvh http://download.fedoraproject.org/pub/epel/7/x86_64/e/epel-release-7-9.noarch.rpm
-# preparations for r-xml
-yum -y install libxml2 libxml2-devel
-yum -y install R
-Rscript -e 'install.packages("XML", repos="https://cran.rstudio.com/")'
-Rscript -e 'install.packages("lattice", repos="https://cran.rstudio.com/")'
+# chmod -R o+x $INSTALL_ROOT
 
-chmod -R o+x $install_root
-
-
-
-#autoinstall script
-
-
-# do not run this script if wildfly already present
-# otherwise this will likely break the installation
-if [ -f $JBOSS7_DIR ]
-then
-	>&2 echo "Aborting $0, wildfly is already configured"	
-	exit 1
-fi
-
+echo
+echo +++++ STEP 2.02 +++++ Wildfly Anpassung | tee -a $LOGFILE
+echo
 # create symlink for fixed configuration paths in i2b2
-ln -s $WILDFLY_HOME $JBOSS7_DIR
+ln -s $WILDFLY_HOME $JBOSS7_DIR | tee -a $LOGFILE
+
 
 if [ ! -d "$DATA_DEST" ]; then 
     mkdir $DATA_DEST
+    chmod -R 777 $DATA_DEST
 fi
-cp -r -f $DATA_HOME/* $DATA_DEST
-cd $DATA_DEST
 
-buildfile=build.properties
-# add some system and build dependent parameters for the ant build
-echo "# system generated properties for ant build" >> $buildfile
-echo "ant.installdata.dir=${DATA_DEST}" >> $buildfile
-echo "i2b2.src.dir=${DATA_DEST}/i2b2_src" >> $buildfile
-echo "packages.dir=${PACKAGES}" >> $buildfile
-echo "install-log.dir=${LOG_DIR}/ant-install" >> $buildfile
-# add wildfly data
-echo "app.base.dir=/opt" >> $buildfile
-echo "jboss.home=${WILDFLY_HOME}" >> $buildfile
-# system dependent postgres sql call
-echo "postgres.users.create=create_POSTGRESQL_users_cent" >> $buildfile
-# change project name TODO needs refining
-echo "db.project=demo" >> $buildfile
-echo "db.project.name=Demo" >> $buildfile
-echo "db.project.uname=i2b2_DEMO" >> $buildfile
-echo "db.hive.id=i2b2demo" >> $buildfile
 
-echo ant scripts
-ant all
 
-### import ontology to postgres
-CDATMPDIR=/var/tmp/cda-ontology
+# up till now, the script can be rerun. but not if it dies while ant is running.
+echo
+echo +++++ STEP 3 +++++ Installation via ANT | tee -a $LOGFILE
+echo
+if [ ! -d "$DATA_DEST" ]; then 
+    mkdir $DATA_DEST
+fi
+if [ -d $WILDFLY_HOME ] && [ -d $WILDFLY_HOME/standalone/deployments/i2b2.war ] ; then
+	# i2b2 is already installed (or at least some part of it. abort! )
+else 
+	cp -r -f $DATA_HOME/* $DATA_DEST
+	cd $DATA_DEST
 
-touch update_sql.log
-echo "update ontologies to ${project.dependencies[1].version}" 2>&1 | tee -a update_sql.log
-# unzip the sql jar 
-unzip $PACKAGES/cda-ontology-${project.dependencies[1].version}.jar -d $CDATMPDIR
-chmod 777 -R $CDATMPDIR
+	buildfile=build.properties
+	# add some system and build dependent parameters for the ant build
+	echo "# system generated properties for ant build" >> $buildfile
+	echo "ant.installdata.dir=${DATA_DEST}" >> $buildfile
+	echo "i2b2.src.dir=${DATA_DEST}/i2b2_src" >> $buildfile
+	echo "packages.dir=${PACKAGES}" >> $buildfile
+	echo "install-log.dir=${LOG_DIR}/ant-install" >> $buildfile
+	# add wildfly data
+	echo "app.base.dir=/opt" >> $buildfile
+	echo "jboss.home=${WILDFLY_HOME}" >> $buildfile
+	# system dependent postgres sql call
+	echo "postgres.users.create=create_POSTGRESQL_users_cent" >> $buildfile
+	# change project name TODO needs refining
+	echo "db.project=demo" >> $buildfile
+	echo "db.project.name=Demo" >> $buildfile
+	echo "db.project.uname=i2b2_DEMO" >> $buildfile
+	echo "db.hive.id=i2b2demo" >> $buildfile
 
-# call sql script files. no console output since spamming
-echo "update metadata " 2>&1 | tee -a update_sql.log
-su - postgres bash -c "psql -d i2b2 -f $CDATMPDIR/sql/meta.sql" 2>&1 >> update_sql.log
-echo "update crcdata " 2>&1 | tee -a update_sql.log
-su - postgres bash -c "psql -d i2b2 -f $CDATMPDIR/sql/data.sql" 2>&1 >> update_sql.log
+	echo ant scripts | tee -a $LOGFILE
+	ant all 2>&1 | tee -a $LOGFILE
 
-# remove temp directory
-rm -r $CDATMPDIR
+	cd $INSTALL_ROOT
+fi
 
-#TODO
-# add apache to autostart
-# add wildfly to autostart
+echo
+echo +++++ STEP 4 +++++ Set Up Wildfly in autostart and start it | tee -a $LOGFILE
+echo
 
-# restart server!
-# apachectl restart
-# /opt/wildfly-9.0.2.Final/bin/standalone.sh -Djboss.http.port=9090 > /opt/aktin/logs/wildfly.log &
 
 ln -s $WILDFLY_HOME /opt/wildfly
 
-if [ ! -f /etc/default/wildfly.conf ]
-then
+if [ ! -f /etc/default/wildfly.conf ]; then
+	echo erstelle /etc/default/wildfly.conf  | tee -a $LOGFILE
+
 	touch /etc/default/wildfly.conf
-
 	#  cp /opt/wildfly/bin/init.d/wildfly.conf /etc/default/wildfly.conf
-
 	echo JBOSS_HOME=\"$WILDFLY_HOME\" >> /etc/default/wildfly.conf
 	echo JBOSS_USER=wildfly >> /etc/default/wildfly.conf
 	echo JBOSS_MODE=standalone >> /etc/default/wildfly.conf
@@ -162,20 +208,25 @@ then
 	echo JBOSS_OPTS=\"-Djboss.http.port=9090 -Djboss.as.management.blocking.timeout=6000\" >> /etc/default/wildfly.conf
 fi
 
-cp /opt/wildfly/bin/init.d/wildfly-init-redhat.sh /etc/init.d/wildfly
+if [ ! -f /etc/init.d/wildfly ] ; then 
+	echo kopiere den Wildfly init skript nach /etc/init.d/wildfly | tee -a $LOGFILE
+	cp /opt/wildfly/bin/init.d/wildfly-init-redhat.sh /etc/init.d/wildfly
+fi
+systemctl daemon-reload 2>&1 | tee -a $LOGFILE
 
 mkdir -p /var/log/wildfly
 
-useradd --system wildfly
+echo erstelle user wildfly  | tee -a $LOGFILE
+useradd --system wildfly 2>&1 | tee -a $LOGFILE
 
-chown -R wildfly:wildfly $WILDFLY_HOME
-chown -R wildfly:wildfly /opt/wildfly
-chown -R wildfly:wildfly /var/log/wildfly
+echo rechte ändern  | tee -a $LOGFILE
+chown -R wildfly:wildfly $WILDFLY_HOME 2>&1 | tee -a $LOGFILE
+chown -R wildfly:wildfly /opt/wildfly 2>&1 | tee -a $LOGFILE
+chown -R wildfly:wildfly /var/log/wildfly 2>&1 | tee -a $LOGFILE
 
-systemctl daemon-reload
-
-# make 0.7 changes with aktin_dwh_07_config.sh
-$install_root/aktin_dwh_07_config.sh
 
 systemctl enable wildfly
 systemctl start wildfly
+
+
+# call update script (step 1 mitnehmen)
